+++ contrib/scrooge/tests/python/pants_test/contrib/scrooge/tasks/test_scrooge_gen.py
@@ -15,7 +15,7 @@ from pants.backend.jvm.targets.scala_library import ScalaLibrary
 from pants.base.address import SyntheticAddress
 from pants.base.build_environment import get_buildroot
 from pants.base.build_file_aliases import BuildFileAliases
-from pants.base.exceptions import TargetDefinitionException, TaskError
+from pants.base.exceptions import TaskError
 from pants.goal.context import Context
 from pants.util.dirutil import safe_rmtree
 from pants_test.base.context_utils import create_options
@@ -42,30 +42,6 @@ class ScroogeGenTest(TaskTestBase):
   def tearDown(self):
     super(ScroogeGenTest, self).tearDown()
     safe_rmtree(self.task_outdir)
-
-  def test_defaults(self):
-    self.add_to_build_file('test_validate', dedent('''
-      java_thrift_library(name='one',
-        sources=[],
-        dependencies=[':one'],
-      )
-    '''))
-    target = self.target('test_validate:one')
-    ScroogeGen._validate([target])
-    self.assertEquals('sync', target.rpc_style)
-    self.assertEquals('thrift', target.compiler)
-    self.assertEquals('java', target.language)
-
-  def test_invalid_value(self):
-    self.add_to_build_file('test_validate', dedent('''
-      java_thrift_library(name='one',
-        sources=[],
-        rpc_style='xyz',
-        dependencies=[':one'],
-      )
-    '''))
-    with pytest.raises(TargetDefinitionException):
-      self.target('test_validate:one')
 
   def test_validate(self):
     self.add_to_build_file('test_validate', dedent('''
